+++ lib/OrganizationsFieldArray/RolesFieldArray.js
@@ -2,6 +2,7 @@ import { useEffect, useState, useRef } from 'react';
 import PropTypes from 'prop-types';
 import { FormattedMessage } from 'react-intl';
 import { Field } from 'react-final-form';
+
 import {
   Button,
   Col,
@@ -12,16 +13,9 @@ import {
   TextArea,
 } from '@folio/stripes/components';
 import { useKiwtFieldArray } from '@k-int/stripes-kint-components';
-import { required } from '../validators';
 
-// custom hook that holds any required value (props/state) from  the previous render cycle via a ref.
-function usePrevious(value) {
-  const ref = useRef();
-  useEffect(() => {
-    ref.current = value;
-  }, [value]);
-  return ref.current;
-}
+import { required } from '../validators';
+import { usePrevious } from '../hooks';
 
 // Utility function to check if two arrays of scalars contain the same items (Order does not count)
 const arraysAreEqual = (a, b) => {
