+++ lib/hooks/index.js
@@ -12,3 +12,4 @@ export { default as useSingleFetchInterfaceCredentials } from './useSingleFetchI
 export { default as useChunkedCQLFetch } from './useChunkedCQLFetch';
 export { default as useChunkedUsers } from './useChunkedUsers';
 export { default as useParallelBatchFetch } from './useParallelBatchFetch';
+export { default as usePrevious } from './usePrevious';
